Swagger refactor to isolate v1 & v2 APIs
The mixed mode aspect of DataAPI is problematic for swagger and violates
many swagger assumptions. Swagger assumes 1 binary with 1 router that serves all API
versions, but with DataAPI v2 the v1 API are not available, and each
server mode instatiates its own router. To isolate v1/v2 API we define
separate swagger instances/packages. The real problematic piece is how
swagger dynamically finds handler endpoints. The only way to get it to
ignore v1 or v2 when build a specific version is to define handler
annotations in a subdir and generate swagger from that directory.