chore: bump default confirmation-depth 0->8 blocks (#285)
* chore: change default confirmation-depth to 5 blocks (from 0 blocks)

This is a very important change. Reorgs of 1 block are quite common and can cause a da cert to be submitted to the rollup inbox with a wrong confirmationBlock! It's important to wait a few blocks for the confirmBatch to have landed onchain before submitting the dacert to the batcher inbox.

* docs: add troubleshooting guide with first section on the batch-hash-mismatch error

* chore: bump default confirmationDepth 5->8