Fix node block stale measure & store duration overflows when test-mode is enabled
If `NODE_ENABLE_TEST_MODE=true` the node will evaluate `NODE_OVERRIDE_BLOCK_STALE_MEASURE` and
`NODE_OVERRIDE_STORE_DURATION_BLOCKS` which default to -1 and get stored in Config as `int64`.
Before evaluating if test mode is enabled and the overrides should be enabled, the override
values were cast to `uint32` to match return type of `tx.GetBlockStaleMeasure`. This resulted
in the default value -1 being cast to `18446744073709551615`.

This changes updates the flag type to `Uint64Flag` with default value of 0

This change also adds the override values to the NodeInfo response if
overrides are enabled. This will allow us to observe if any nodes are explicitly load shedding /
withholding data via overrides.