ci: separate holesky-test into its own job (#325)
* ci: separate holesky-test into its own test

Also lower its timeout to 30mins
Holesky is broken so this test is flaky. Makes sense to sepraate it into its own job so that we can rerun only this test when it fails.

* ci: remove 30 min timeout on workflow in favor of that in makefile

wasnt seeing logs because ci job was getting cancelled before test failed, and didn't have -v flag enabled.
So enabled the flag and removed the timeout from ci in case