chore: better errors and docs for batch-hash-mismatch error (#322)
* chore: create structured error for batch-hash-mismatch error

This permits handling it. Realized while implementing this that there is currently no good way to deal with the error however... but decided to still commit it for better documentation of why this is not possible.

* docs(troubleshooting_v1.md): expand on batch-hash-mismatch error documentation

Explain that this is only possible on op-stack.

* docs(troubleshooting_v1): add explanation that batch-hash-mismatch also affects nitro l3s

* docs(batch-hash-mismatch): remove TODO since this error is not solvable

Added explanation for why not.

* docs(batch-hash-mismatch): update old incorrect documentation for BatchMetadataHashMismatchError

Not actually possible to handle this error. Update the explanation, and point to new protobuf docs

* docs: fix typo

* fix: move errors.go to verify/v1 (it was left alone and breaking build after merge master)