docs: cleanup readme by separating users from contributors (#330)
* docs: cleanup readme by separating users from contributors

first time contributors were lost and didn't understand how to setup proxy and interact with it.

Added a user guide to the readme which gives a quick start example of spinning up proxy in memstore mode, dispersing and retrieving a payload.

Also refactored a bunch of the sections and TOC to make the whole readme easier to parse.

* docs(README): rephrase GET route documentation

* docs(README): break up long lines