chore: isolate keccak commitment POST route (#386)
* chore: remove unused commitment functions/types

* chore(makefile): install mockgen in go-gen-mocks target

Was having local issues because I had the uber mockgen installed locally instead of golang/mock/mockgen

* chore: get rid of commitmentMeta

We had commitmentMeta which consisted of commitmentMode + certVersion. This didn't really make sense to me as the certVersion should go with the cert.

So created a new struct VersionedCert which contains the certVersion and serializedCert. And then functions take this and a commitmentMode, as opposed to the serializedCert and commitmentMeta.

I believe this is a simplifying change. Its also forward looking because we will need to allow for new cert types. I plan to follow up with a cert/rbn parsing library, which will further simplify some of the logic in proxy.

* docs: fix typo

* chore: replace golangci deprecated tenv with usetesting

* chore: add constructor for GETError and POSTError

This prevents errors where we forget to instantiate some fields because we don't have the linter exhaustruct turned on

* chore: add ByteToEigenDACertVersion parsing function

* Revert "chore: replace golangci deprecated tenv with usetesting"

was only an issue because I use a newer version locally, but our CI
doesn't support usetesting (its still on golanci v1.61

This reverts commit 370e25edd16983334fb8a4438d6dc117cece8458.

* docs: add comment better explaining CertV1 is for blazar

* style: EigenDAVersionedCert.CertVersion -> Version

* chore: rename interface GeneratedKeyStore -> EigenDAStore

* style: fix error strings in handlers (commitment->versionedCert)

* Update commitments/op.go

Co-authored-by: ethenotethan <ethen@eigenlabs.org>

* docs: remove TODO comment

* fix: error != bug

* style: fix error string that was causing test error comparison errors

* chore: simplify commitment types and methods

* chore: move certs and commitments under common/types

* style: rename EigenDAVersionedCert -> VersionedCert

We know certs in eigenda-proxy are eigenda certs, no need for the extra length

* docs: add docstring to EncodeCommitment

* style: EigenDADALayerByte -> EigenDALayerByte

* style: delete useless comment for NewStandardCommitment constructor

* chore: isolate op keccak PUT route

keccak has a very different semantic from standard and op generic commitment modes... because it writes to s3 and doesn't need to deal with eigenda certs.

It didn't make sense for it to be behind the same store interface.

There is still some cleaning up to do (left some TODO comments in the code), but this is at least a start.

* style: return hex instead of byte commitment in POST error

* style: clean up Keccak256KeyValueMismatchError + add docs

* style: change error wrapping msg convention in manager.go

---------

Co-authored-by: ethenotethan <ethen@eigenlabs.org>