chore: repo organization refactor (#396)
* chore: move verify package next to eigenda v1

no reason for this to be top level now that we have v1 and v2 in same repo, and the verify package is only used for v1.

* chore: move all test related stuff under test/

* refactor: server manages its own config/flags

appconfig also refactored to have an explicit high-level server config, not hidden inside EigenDA config

To fix the circular issue that happened from moving flags/config to server package, got rid of the BuildAndStartProxyServer constructor and included that as part of the main entrypoint/testbuild functions. I personally think the server should only be a server that gets injected a store, it shouldn't know how to build a store.

* refactor: storeManagerBuilder

Storage configs and naming were convoluted and hard to follow. Wanted to make the builder take the whole ProxyConfig as input instead of all its individual fields. This small task ended up bubbling to this huge changeset...

Had to move the "proxyConfig" to the store in order to prevent import cycles. Made me realize the store "config" was a weird amalgation of store + secondary storage config. So broke that up, and created a separate builder package.

Also got rid of the StoreManagerBuilder class in favor of just functions. Find this cleaner personally, though I admit it wasn't strictly necessary.

* docs: remove outdated comment

* style: rename ReadCLIConfig -> ReadAppConfig

* style: rename proxyConfig -> storeBuilderConfig

* style: consistent store naming (no more storage)

* ci: fix test-e2e-sepolia command