chore: add basic CODEOWNERS file
Want to force any client changes to be reviewed by at least one member of the integration team, to keep us in the loop.

Also added pakim as owner of contracts, and anup as owner of docs/audits.

We should probably define some smaller teams like devops/infra, etc. and assign these teams to specifc parts of the codebase.