feat: introduce error middleware (#400)
* feat: introduce error middleware

We were previously returning 200s on requests that should have returned 4xx/5xx errors.

Adding an error middleware to consistently deal with errors. This is a wip as there are still unstructured errors that are not being probably filtered by the error middleware right now. But we can use this framework to gradually add them and continue cleaning up our error handling (which was in a pretty bad shape).

Also refactored our middlewares to chain logging -> metrics -> errors. I might have created regressions on the metrics and logging side.. as I also cleaned up the way request mode and cert versions were propagated (previously via errors... which is just clearly wrong). These are now being propagated via a request context... open to changing if there is a better way.

* fix: lint issues

* fix: wrap all routes in error middleware

the helper routes like set-dispersal-backend were not wrapped in error middleware. Thankfully Austin added a unit test that caught this. I'm sure there's similar regressions like this hiding... we need better test coverage for middleware stuff.

* style: key -> Key

* style(4xx): errors in same order

* style: eg. -> e.g.

* refactor: requestContext used in middlewares

Also refactored handlers to separate into 2 files main cert-related routes and helper routes. In this process decided to completely remove use of middlewares for helper routes. They are supposed to be simple anyways, so can handle their own logging/errors without too much clutter.

The middlewares are currently hardcoded to only work for cert routes. We could try to expand them and make them work for all routes in the future, but its not even clear how much benefit we get out of that.

* test: add request_context test

Make sure metris middleware can read certVersion set by the handler properly.

* chore: get rid of POSTError structured error

It was relaying useless information (mode) only. I don't even think its needed, and even if it is, there are better ways to do this, which we need to keep looking for by continuing to refactor this code.