chore: use mise to track deps (#1604)
* chore: use mise to track dependencies

there's a bunch of implicit tooling dependencies in the repo, scripts, and ci. Introducing mise to keep track of these dependencies and make them reproducible locally and in ci. Only added go and golangci-lint for now, but we can keep adding to this.

* chore: add golangci config file

Currently doesn't turn on any non-default linters, but it's ready if we ever want to (for example, we are using the deprecated golint linter, which revive supercedes)

* chore(makefile): cleanup lint/fmt commands

* ci(lint): update golangci-lint to use make lint/fmt-check

* ci: replace setup-go action with mise-action

this installs go using mise, for reproducible local/ci tooling

* ci: fix checkout/mise ordering

checkout repo should be done before installing deps with mise, because repo contains mise.toml file which is used by mise.

* ci: increase golangci-lint timeout to 5m

default of 1m was timing out in ci

* docs: add mise section in README

* chore: fix make fmt command

* chore: make fmt

* style(ci): remove old commented out code

* style: make fmt