fix(make): fmt-check
Previously the gofmt -l command was running through every single go file in subdirs, including contract submodules. It was erroring on a non-formatted file from eigenlayer-middleware I believe, or some other contract dep libs.

This didn't happen in CI because the ci that runs this doesn't checkout recursively the repo. But it happened locally and was annoying. So this fixes it.