chore: remove bytecode from bindings so they only contain abi (#1636)
* chore: change contracts compile.sh to generate bindings with only abi

We remove the bytecode from bindings. This was causing all sorts of CI issues (see https://github.com/Layr-Labs/eigenda/pull/1635).

I believe we weren't deploying any contracts from golang code, so this should be fine. Anyone who does want to do this can create their own "deployable-bindings" (we could also create these and check them in to the repo if needed...)

* chore: make compile-contracts

Regenerate bindings to remove bytecode, so they only contain abi now.

* ci(contracts): turn on bindings diff check

This was turned on because ci was choking on some bytecode changes that we don't understand. Now that are bindings only contain abi, this test should be turned on again.

* fix: contracts/compile.sh script