feat: return HTTP teapot error for invalid certs (#406)
* feat: convert CheckDACert structured errors to HTTP 418

The bump to recent commit has broken compilation. A lot of the logic needs to be updated to reflect the new interface changes. Those changes have already been implemented in https://github.com/Layr-Labs/eigenda-proxy/pull/384, so will wait for that to merge and rebase on top.

* docs: add TODO comment for dealing with cert rlp decoding error

* docs: better docs for v2 Verify method

* chore: write 418 error status codes to body as json

* fix: is418 errors.As bug (need poitner to error)

* fix: 418 handling in error middleware

* fix: error middleware writes CertVerification status code to teapot body

* test: update op_contract_test to test op client

This relies on https://github.com/Layr-Labs/optimism/pull/49
We need to merge that PR and then update our go.mod replace directive

* test: add error middleware unit tests

* fix: lint

* style: fix error middleware certVerificationFailed error path

* fix: error middleware rbnRecencyCheck panic when marshal fails

* chore: update op commit

After merging https://github.com/Layr-Labs/optimism/pull/49. commit is different (sqaush commit) but code is exactly the same, so it shouldn't make a functional difference.

* feat: define new CertVerificationFailedError status codes

These are used for different cert verification error cases. RBNRecencyCheck is no loner a standalone error case, it was moved to be part of the CertVerificationFailedError. To achieve this its status code was changed from -1 to max(uint8).

* style: fix lint errors

* chore: make format

* docs: fix TODO comment for ParsingError