fix: relative paths in test/v2/config files
Before relative paths wrt test files were used in the config file (for SRSPaths), which only worked for the tests, but not for running the load-generator.

I made the paths relative to the config file itself, so now they work for both tests and load-generators. I think it also makes understanding the config files easier (otherwise there's some mysterious relative path which doesn't make sense without knowing that the paths are meant to be interpreted wrt where the tests are located).