perf(meterer): add accountLock to prevent dynamo tx fails
Single account submitting multiple dispersal requests causes TxWrites failures. This causes an error to be returned to the client, and the client has to resubmit the blob, which is very inefficient when submitting large blobs.

2 solutions:
1. use a lock (this PR)
2. retry the dynamo atomic update tx a few times before returning error to client

We will try the lock approach first, test it agains the load generator, and if performance drops significantly, then consider the second approach or alternatives.