fix: node shouldnt require eigenda directory
Preprod Validator was seeing 2025/07/16 22:21:34 application failed: failed to create address directory reader: address directory must be a valid hex address:
When using the old config. EigenDADirectory was added, but should be backward compatible and just seamlessly fallback to using the provided EigenDAServiceManager passed in as argument when Directory is not present.

We will later deprecate and remove the other flags but for now we want the flags to be backwards compatible.