chore(cert_verifier): panic on invalid cert version
instead of returning a 4xx like error. This is a bug in our implementation so shouldnt return a normal error imo.

From the pov of the proxy, the panic will result in a 5xx being returned to the client (say hokulea), so the result will be the same. But I just personally (and talked about this with bowen) feel that the panic is cleaner here, because this is clearly a bug in the client code, and not a normal 4xx error.