chore(proxy-clients): new sub go module
We were still using the eigenda-proxy/clients package. Now moving this package over to monorepo.

This is also a good chance to cleanup the weird godoc behavior that won't listen to our retract directive in go.mod to remove the 1.0.1 version.

So with this new package our client docs will at least point to the most up-to-date examples.

TODO: publish this new module as v0.2.1 so that it lands on pkg.dev (using same version as the current one for the eigenda-proxy/clients pkg)

TODO: will need to update our eigencloud docs to point to this new module.