feat(proxy): implement new teapot statuscode scheme (#1736)
* feat(wip): implement new teapot statuscode scheme

* feat: implement derivation errors which return teapots

* style: fix lint

* fix: middleware teapot error_test

* chore: error wrapping nit

* fix: error cases in encodedPayload.decode

Not super sure why these were not being tested. Need to check with Austin whether this is a real mistake or I'm missing something here.

* feat: finish implementing new teapot error scheme

Also moved the errors to client so that the blob decoding error can be returned without causing an import cycle between clients and proxy.

* test: add blob decoding test to relay_payload_retriever

* style: remove print stmts

was using them for debugging

* docs: clarify error log in retriever clients

* style: fix typo on coeff -> in coeff

* docs: malicious blob -> wrongly encoded blob

* style: c -> cert

* style: pipeline -> pipelines

* docs: move Validate() closer to sentinel errors

Also better comment describing the magic error codes (1-4)

* test(middleware): handle all 4 derivation error cases

* docs: document todo better in op_contract_test

* chore(cert_verifier): panic on invalid cert version

instead of returning a 4xx like error. This is a bug in our implementation so shouldnt return a normal error imo.

From the pov of the proxy, the panic will result in a 5xx being returned to the client (say hokulea), so the result will be the same. But I just personally (and talked about this with bowen) feel that the panic is cleaner here, because this is clearly a bug in the client code, and not a normal 4xx error.

* test(op): update op version and fix contract test

* test(proxy-op): clearer test case name

removed the prob llm generated name that was unclear

* test(op-proxy): add cert related tests

Only missing blob decoding tests