chore(proxy): simplify .env example files to a single one
Having multiple example files is a source of bug and confusion. its also hard to maintain. Consolidating into a single .env.example file. People should figure out how to populate it for their own network anyways.

Also removed all the config values that have sane defaults, and only left the ones that are mandatory to set. This removes a lot of cognitive load on our users, and will be a helpful simplification I believe.