chore(proxy): deprecate fallback reads in favor of cache
Fallback read behavior (read from cache target only after trying to read from EigenDA) was afaiu not used by anyone. Opting to clean up the code and keep the configuration simpler.

I don't currently see a reason why a user wouldn't want to read from the cache before EigenDA, since if the data is there, it will clearly be read faster and reduce latency and global network costs.

Reading from EigenDA validators should be the last bastion since it puts a lot of stress on the network. It thus makes more sense to have a layer of caches (including reading from the eigenda relays) in front of the validators, not after them.