feat(tools): new eigenda contract addresses reading cli (#1804)
* wip

* feat(directory_reader): add new GetAllAddresses method

This is used to fetch all addresses from a given EigenDADirectory

* chore(proxy_common): small improvements to network utils

* feat(tools): new contract discovery cli

Given an eth-rpc it prints a table, json, or csv output of all contract names and their address.

* style: remove unused ContractInfo struct

* style: bi -> buildInfo

* docs: clearer networkflag usage string

* style: define global validFormats slice

* refactor: pull redundant code out of if/else

* style: more succinct error msgs

* style: delete useless comment

* style: name -> contract_name

* style(directory_reader): fix error msgs

* fix(directory_reader): add ctx to all fcts

Unfortunately a bunch of places in the codebases are not meant to inject proper contexts... I won't go and fix all of this in this PR. Left some TODO comments.

* fix: remove dangerous DefaultEigenDANetworkFromChainID fct

* refactor: discovery flags

We should do this for every flags in every service... much cleaner!

* fix: error string

* refactor: make network flag required

This simplifies the code by quite a lot. Also forces user to pass in the network that they actually want (instead of having too many implicit choices made for him). I think overall its a good simplification of the UX.

Also will be very useful as we install new subcommands into this CLI, since they will all share the network.