feat(coretypes): add MaliciousOperatorsError
Started from this discussion: https://github.com/Layr-Labs/eigenda/pull/1841#discussion_r2266759587
We realized that we need to differentiate errors that are the result of malicious operators, since these should eventually lead to operators getting slashed.

This PR adds the MaliciousOperatorsError type, and the first error of this kind, for when the dacert's blob length is not a power of 2.