fix: initOnceGrpc function bugs
These functions had a bug where the error was not stored, and hence would not be re-returned on future calls.

Furthermore, realized the function itself is completely unnecessary because creating a new grpc client doesn't actually make any IO calls. Those are only done the first time an rpc is submitted. So this logic was not needed at all. Removed it.