feat: add grpc_recovery middleware on every server
This middleware will recover from any panics in any goroutine serving client requests, and instead of crashing the entire server, return a grpc INTERNAL error to that user.

This behavior is done by default on http servers, but for some reason is not default on grpc servers.

We have already seen our encoders crashing for a single request that was missing some srs table. This would have prevented those requests from crashing the entire encoder server.