refactor: eigendacert interface
We previously had 2 separate EigenDACert and RetrievableEigenDACert interface. I'm still not 100% sure what's the best way to do this but this became very clunky to use when trying to refactor our payload retriever clients to also do verification, since cert verifier takes an EigenDACert instead of a RetrievableEigenDACert.

The main thing I was trying to achieve is an ADT, so I merged both interfaces, which added the private method to our EigenDACert interface. This way no one else can implement this interface, and we are free to break the API in the future if we ever need to change our cert representation or the way this interface is handled/represented.