refactor(proxy): separate eigenda_mgr getEigenDAV1/V2
The previous coupling made it very hard to move the v2 functionality into a higher-level verifyingPayloadRetriever client that verifies the cert, the rbn recency check, and proceeds to retrieve from all configured backends.

Writing this first refactor PR to clearly separate v1 from v2 behavior in proxy, in order to enable this next refactor. This duplicates the secondary storage logic in both getEigenDAV1 and getEigenDAV2, but that both:
1. made me find some bug with the verify fct for v2 (see inline comments)
2. will soon be refactored for the v2 path into a higher-level client anyways

Also this refactor is another instance of one of my favorite programming mottos: https://matklad.github.io/2023/11/15/push-ifs-up-and-fors-down.html