fix(CertVerifier): add error codes so fct doesnt revert
We use Steel which is not able to prove functions that revert (see https://github.com/risc0/risc0-ethereum/issues/438), so we need to turn all reverts into error codes.

This commit is a first step in that direction, adding error codes for abi.decoding faling, and for the call to verifySignature failing.

This endeavor to reach no reverts for any invalid cert (that needs to be proven so using Steel in order to be discarded from a rollup's pipeline) will still require a lot more careful work and audit of our library and downstream dependencies.