fix: better srs and encoding error handling (#1971)
* fix(disperser): better error handling for commitment mismatches

Currently only a 400 is being returned when commitments between client and disperser don't match. This adds better error handling to say which commitment failed to match (commitment or length commitment, etc)

* downgrade Info log to Debug log

Its super noisy in inabox

* wrap errors in encoding/kzg

* docs(inabox): fix readme command to use kzgpad directly

* fix(inabox): wait-for script more informative echo

when operation times out, also print arguments of the call so we know which process failed to start.

* docs(inabox-readme): install instructions for kzgpad