feat(srs): new package with embedded g2 powerOf2 points
The power of 2 points are currently loaded dynamically one at a time from a 1.8KiB file whenver needed. This doesn't make sense, and its a lot more hassle, and the powerOf2 file needs to be carried around. 1.8KiB is nothing, so better to just embed in all binaries that need it and call it a day. Multiframe verification will also run faster thanks to this.