refactor(encoding): introduce parametrized_verifier (#1972)
* refactor(encoding): cleanup some unused functions/variables

Also made some public methods private, and created parametrized_verifier the same way we have a parametrized_prover.

* delete encoding/test/main.go

Seems like an old piece of code which was used during development. Hasnt been touched for 2 years and not being run in CI or anywhere afaiu.

* delete kzg struct

Was just a useless thin wrapper around fftsettings and srs, which we have access to already in all places individually.

* remove Prover comment