refactor(srs): fix SRSOrder to 2^28
We should never have changed the SRSOrder. The real srs order is 2^28 for bn254, and we should always use it that way. Instead of doing hacks in tests/inabox of specifying a lower srs order, we should instead be passing the g2 trailing file like we now do elsewhere.

This commit does that. There are still some broken tests to fix, due to needing to read trailing G1 points. I will open a new PR to embed the 28 G1 points that we need to verify length proofs, like we did for the G2PowerOf2 embedded points.