feat(srs): embed G1PowerOf2SRS and stop reading individual points
Previous code needed to read trailing points of the G1 file (G1file[2^28 - 2^i]). By enforcing that all blobs are powers of 2, and fixed the SRS order, we really only ever need to read 28 G1 points. We hardcode those and embed them into the binaries.

One weird thing I realized is that the prover doesn't check and/or enforce that the data is receives is a power of 2 length. And the tests were not enforcing that. Changes the ToFrArray to pad the data it encodes to the next power of 2. That loses some performance since we might be sending some 0s over the wire, but at least it allows the other optimization of not needing to read the entire G1 file like operators do today. Also we can always change that function and only pad on the disperser if ever needed.