refactor(prover-v2): cleanup unused fcts (#2057)
* chore(prover): delete unused functions + make private fcts that should be

* refactor: move decode method to parametrized_prover v2 file

no idea why it was in its own file...

* docs: move comment from makefile to fuzz file

* chore: move prover_fuzz_test test to prover_test

* test: add FuzzOnlySystematic fuzz test to makefile target

* refactor: make ParametrizedProver field private

* refactor: ParametrizedProver only takes 2 fields it needs from kzgConfig

* refactor(prover-v2): delete EncodeAndProve

There is no place in the v2 system where we both want to compute commitments and generate frames anymore. Commitments are done client side or on the apiserver, and frames are computed on the encoder. So we no longer need this function.

* refactor: delete unused ParametrizedProver.EncodeBytes

This was only used in tests. To update tests had to change a bunch of places which were taking []Frames instead of []*Frames. Our codebase is very not consistent with this. I think there are probably a lot of other places that we should all change to be consistent.

* chore(prover-v2): delete Decode fcts

These were only used in tests. verifier also has the exact same functions which I switched to using in tests. Not even sure verifier should have these functions but we'll cross that bridge at a later time.

* refactor(prover-v2): delete Encode

Following last few commits, v2 prover NEVER has to both generate frames AND commitments. Those are generated in different places (client/apiserver vs encoder). We should eventually just get rid of the commitment generation from the prover and make that its own separate struct/service.

* docs(prover-v2): document TODO to split commitment from prover

* style: fix lint

* style: fix typo computeMultiproofNumWorker

* chore: move FuzzOnlySystematic test to root Makefile

I had put it in proxy makefile by mistake

* chore(gitignore): add **/testdata and **/SRSTable

* test: fix v2 prover_test FuzzOnlySystematic

* fix: wrong EncodingParams in getAllPrecomputedSrsMap

* fix icicle