refactor(inabox): cleanup inabox deployer + makefile (#2070)
* refactor(inabox): cleanup inabox deployer + makefile

Simplify inabox deployment. Everyone is always confused as to how to start it. There's too many commands. We always want to start all of it, so no need for all the subcommands that start portions of it. Got rid of all of that and only left with simple makefile commands:
1. make new-anvil (prob can find a better name...)
2. make start-infra
3. make start-services (calls bin.sh)

* fix tests to use testconfig-anvil.yaml

* delete testconfig-anvil-nograph.yaml

Don't think there's ever a reason to run inabox without a subgraph, it probably doesn't even work (we need to read g2 points).

* test: fix thegraph test

Using the testconfig-anvil.yaml file instead of nochurner one (which was deleted), so number of operators expected by indexer is 3 instead of 4, because of churner settings.

* fix lint

* fix graph integration test: bring back nochurner yaml file

Cry. didn't realize thegraph test doesn't spin up churner. Anyways we're getting rid of offchain churner soon so we'll be able to get rid of this file then.