feat(proxy): v2 read-only mode when no privateKey
Proxy was previously forced to be provided with a signerPrivateKey. In order to start in read-only mode, we were passing a bogus privateKey (see our .env.example).

Changed our config/flags/examples to allow not passing a SignerPrivateKey and explicitly marking that flag as optional.

This is a request from Celo, who found a few of their operators were confused and kept asking about this.

Note that this change was only made for EigenDA v2 clients, not V1.