perf: introduce crypto/encoding benchmark suite (#2154)
* perf: introduce crypto/encoding benchmark suite

Documents and writes benchmarks for all of the low-level primitives that we use (fft+msm on Fr/G1/G2).

Also adds high-level benchmarks for all of the higher-level operations that we perform (blob encoding, commitment generation, chunk+proofs generation).

Finally, adds a icicle benchmark suite so that we can properly benchmark GPU operations and how much improvement they actually bring over parallelized (which our FFTs are currently not!) implementations of the same operations.

* fix lint + copilot comments

* add //go:build icicle tag

* rename benchmark packages to have _test suffix

* add README

* BenchmarkBlobToChunksEncoding also tests blobPower 21

* witout -> without

* add makefile target to run golang_benchmarks in encoding/bench

* make BenchmarkMultiproofFrameGeneration bench multiple blob sizes