refactor(prover-v2): pull encoder out of parametrizedProver (#2198)
* refactor(prover-v2): pull encoder out of parametrizedProver

There are a few reasons for doing this:
First is that bowen found an optimization where encoding and proving will now require using different parameters, so it doesn't make sense for them to both be parametrized the same way.
Second ParametrizedProver just held too many things, and is hard to follow the dependency of things. Namely, this made me realize that the encoder held by the parametrizedProver was a normal encoder (which itself holds a map of parametrizedEncoder), so it feels very weird to have it held inside the parametrizedProver.
Third, I'd eventually like to completely pull the encoder out of the prover, and create a small external wrapper struct to combine them to generate a frame (don't think this should be the functionality of the prover, which should focus on generating proofs only).

* fix TestEncodeDecodeFrame_AreInverses

* nit: rename encodeCRes more consistently

* fix wrong indices