fix(encoder-v2-icicle): backend initialization bug
Introduced this bug in https://github.com/Layr-Labs/eigenda/pull/2197, when I refactored the icicle backend initialization logic. I thought the g1SRS was not needed, but it turns out it is.

We don't have any CI that runs gpu code, so this was not caught. Caught it while benchmarking on ec2 today. Bug was causing a divide by zero panic during msm, because the SRS table was not properly loaded.

Personally think the icicle backend config and initialization code is full of footguns, so plan to refactor it in a future PR.