perf(rs-icicle): 20x speedup on GPU ntt (used for chunk generation) (#2214)
* refactor(encoder-v2): cleanup parametrized encoder constructor logic

Create backend package to host gnark and icicle implementations in a self-contained.

* fix BenchmarkMultiproofFrameGeneration

* feat: parallelize icicle Fr conversion utils

This speeds up the gpu RS fft by a lot (seeing ~2x speedup on g6.xlarge which only has 2 cpu cores).

* test: add BenchmarkRSBackendIcicle

This isolates the part of chunk generation that is on GPU, and compares cpu vs gpu performance.

* improve benchmark

* perf: ~30x icicle rs extend_poly throughput

* remove trace output from benchmark_eigenda_test.go

* revert icicle utils.go changes

We don't need them with the new implementation. conversion is done on gpu.

* make fmt

* add back gpuLock to ExtendPolyEval

We need a better solution than this but this is currently needed to prevent panic.

* fix benchmark fr lengths

Was benchmarking with too long blobs... so actually achieved speedup will be less than 20x.... my bad.