perf(proofs-gpu): 400ms speedup for multiproof generation (#2220)
* refactor(encoder-v2): cleanup parametrized encoder constructor logic

Create backend package to host gnark and icicle implementations in a self-contained.

* fix BenchmarkMultiproofFrameGeneration

* feat: parallelize icicle Fr conversion utils

This speeds up the gpu RS fft by a lot (seeing ~2x speedup on g6.xlarge which only has 2 cpu cores).

* test: add BenchmarkRSBackendIcicle

This isolates the part of chunk generation that is on GPU, and compares cpu vs gpu performance.

* improve benchmark

* perf: ~30x icicle rs extend_poly throughput

* remove trace output from benchmark_eigenda_test.go

* revert icicle utils.go changes

We don't need them with the new implementation. conversion is done on gpu.

* make fmt

* add back gpuLock to ExtendPolyEval

We need a better solution than this but this is currently needed to prevent panic.

* fix benchmark fr lengths

Was benchmarking with too long blobs... so actually achieved speedup will be less than 20x.... my bad.

* bench: make proof benchmark work for icicle

* fix: remove validateInput in GetProofs()

It was doing the wrong check by checking that we have enough SrsNumberToLoad points, which is wrong because we can have loaded the SRS tables directly without having loaded any srs points directly.

Realized that there's some cleanup that needs to be done in the prover config and srs points loading... but should be done in future PR.

* chore: make gnarkAffineToIcicleAffine private

* perf(icicle-multiproof-backend): remove unnecessary fr conversion

Same trick that was applied in https://github.com/Layr-Labs/eigenda/pull/2214. icicle functions are now generic and can work with gnark's montgomery fr.Element points directly, so we save a few hundred ms by not converting them on CPU.