test(rust): add first integration test (#2280)
* test(rust): add first integration test

This was a first exploration/playing around with the code that eiger wrote. Pretty happy with the result. There's a few API  kinks that I noticed and would like to iron out. Left some TODO in the integration test.

Currently only supports Sepolia. Plan to add support for inabox eventually so that we can run this in CI.

* fix lint issues

* consolidate rust-ci lint checks into a make command

* fix deny.toml missing licenses

* fix ci lint

* update to latest certverifier