feat(proxy): embed srs points into built binaries
This removes the need for loading SRS points from the filesystem, which makes it easier to distribute the proxy binary as a self-contained artifact. Previously the SRS points were packaged in a docker container, but the rust client may need to start a proxy binary without docker, hence this change to embed the SRS points directly into the binary.