chore(rust): cleanup extractor interfaces and add docs (#2405)
* chore(rust): cleanup extractor APIs + add docs

All the storage_keys functions were taking a cert as input, probably to be uniform. This however hides which of the storage slots are actually cert-dependent vs not. I believe it is cleaner to remove the dependency when not needed, so making the change here.

Also added a diagram of the contract dependencies and storage slots the eigenda-ethereum crate extracts.

* make clippy happy

* remove no-longer-present args from fn docs